requirements.txt
pandas==2.3.3
numpy==2.4.3
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
xgboost==3.2.0
shap==0.51.0
imbalanced-learn==0.14.1
optuna==4.8.0
joblib==1.5.3
flask==3.1.3
streamlit==1.55.0
mlflow==3.10.1
pyyaml==6.0.3
pytest==9.0.2